smoothing factorcasting fix
git-svn-id: svn+ssh://143.225.152.8/home/pasquale/SVN/signal-analyzer/branches/v2-development@39 bcba4c40-192e-4161-9d92-76493bb20e91

+++ src/main/java/com/dmp/signalanalyzer/apps/Main.java
@@ -83,9 +83,9 @@ public class Main {
       filterConfiguration.set("step", Double.valueOf(step));
 
       if (clm.getArguments().containsKey(CommandLineOption.smoothingFactor.name())) {
-         filterConfiguration.set("smoothingFactor", Double.valueOf((String) clm.getArguments().get(CommandLineOption.smoothingFactor.name())));
+         filterConfiguration.set("smoothingFactor", Float.valueOf((String) clm.getArguments().get(CommandLineOption.smoothingFactor.name())));
       } else {
-         filterConfiguration.set("smoothingFactor", Double.valueOf(configurationManager.getSmoothingFactor()));
+         filterConfiguration.set("smoothingFactor", Float.valueOf(configurationManager.getSmoothingFactor()));
       }
 
       //  start analysing signal
